Refactor and enhance codebase for maintainability
- Standardized private field naming conventions (e.g., `init` → `_init`).
- Organized utility classes and enums under `FB2Toolbox.Utilities`.
- Removed unused `using` directives and reformatted code for clarity.
- Refactored `FileUtils` and `FB2File` for modularity and robustness.
- Updated `app.config` with cleaner formatting and updated profiles.
- Adjusted UI elements in `MainForm.resx` for consistency.
- Updated test cases to reflect refactored variable names and paths.
- Removed deprecated `FileUtils.cs` and migrated logic to utilities.
- Enhanced metadata handling and schema validation in `FB2File`.
- Improved localization, exception handling, and logging.
- Updated `AssemblyVersion` to `1.9.0.0`.

These changes improve readability, modularity, and user experience.
