+++ src/FB2Toolbox/Entities/FB2Book.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Text;
 
 namespace FB2Toolbox.Entities
 {
-    public class FB2Book: IBook
+    public class FB2Book : IBook
     {
         #region IComparable Members
         public int CompareTo(object obj)
