+++ src/FB2Toolbox/Extensions.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Text;
+﻿using System.Reflection;
 using System.Windows.Forms;
 
 namespace FB2Toolbox
