+++ src/FB2Toolbox/FB2Toolbox.csproj
@@ -100,7 +100,14 @@
     <Compile Include="Config.cs" />
     <Compile Include="Entities\FB2Book.cs" />
     <Compile Include="Entities\IBook.cs" />
-    <Compile Include="FileUtils.cs" />
+    <Compile Include="Utilities\DescriptionElements.cs" />
+    <Compile Include="Utilities\FB2EncoderFallbackBuffer.cs" />
+    <Compile Include="Utilities\FB2Metadata.cs" />
+    <Compile Include="Utilities\FBEncoderFallback.cs" />
+    <Compile Include="Utilities\FileMetadata.cs" />
+    <Compile Include="Utilities\FileOperationResult.cs" />
+    <Compile Include="Utilities\FileProperties.cs" />
+    <Compile Include="Utilities\FB2File.cs" />
     <Compile Include="GuidForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -124,6 +131,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs" />
+    <Compile Include="Utilities\NativeMethods.cs" />
     <EmbeddedResource Include="ChangeProperties.resx">
       <DependentUpon>ChangeProperties.cs</DependentUpon>
     </EmbeddedResource>
@@ -132,6 +140,7 @@
     </EmbeddedResource>
     <EmbeddedResource Include="MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
+      <SubType>Designer</SubType>
     </EmbeddedResource>
     <EmbeddedResource Include="MainForm.ru.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
