+++ src/FB2Toolbox/Utilities/DescriptionElements.cs
@@ -1,0 +1,28 @@
+﻿namespace FB2Toolbox.Utilities
+{
+    public enum DescriptionElements
+    {
+        #region DescriptionElements elements
+        Genre,
+        Title,
+        SequenceName,
+        SequenceNr,
+        Version,
+
+        AuthorFirstName,
+        AuthorMiddleName,
+        AuthorLastName,
+        AuthorFirstName1,
+        AuthorMiddleName1,
+        AuthorLastName1,
+
+        TranslatorFirstName,
+        TranslatorMiddleName,
+        TranslatorLastName,
+        TranslatorFirstName1,
+        TranslatorMiddleName1,
+        TranslatorLastName1,
+        Lang
+        #endregion
+    }
+}
