+++ src/FB2Toolbox/Utilities/FileOperationResult.cs
@@ -1,0 +1,9 @@
+﻿namespace FB2Toolbox.Utilities
+{
+    public class FileOperationResult
+    {
+        public string NewFullName { get; set; }
+        public string NewFileName { get; set; }
+        public bool Skipped { get; set; }
+    }
+}
