+++ src/FB2Toolbox/Utilities/FileProperties.cs
@@ -1,0 +1,21 @@
+﻿namespace FB2Toolbox.Utilities
+{
+    public class FileProperties
+    {
+        public bool AuthorFirstNameChange { get; set; }
+        public string AuthorFirstName { get; set; }
+        public bool AuthorLastNameChange { get; set; }
+        public string AuthorLastName { get; set; }
+        public bool AuthorMiddleNameChange { get; set; }
+        public string AuthorMiddleName { get; set; }
+        public bool GengeChange { get; set; }
+        public string Genre { get; set; }
+        public string GenreTitle { get; set; }
+        public bool SeriesChange { get; set; }
+        public string Series { get; set; }
+        public bool NumberChange { get; set; }
+        public string Number { get; set; }
+        public bool TitleChange { get; set; }
+        public string Title { get; set; }
+    }
+}
